Add files via upload
separated static classes into separate files
